Add missing fields in emergency_project
