+++ api/drf_views.py
@@ -119,7 +119,26 @@ from .serializers import (
     GoHistoricalSerializer,
     CountryOfFieldReportToReviewSerializer,
 )
-from api.filter_set import UserFilterSet
+from api.filter_set import (
+    UserFilterSet,
+    CountryFilter,
+    CountryFilterRMD,
+    DistrictRMDFilter,
+    RegionKeyFigureFilter,
+    CountryKeyFigureFilter,
+    RegionSnippetFilter,
+    CountrySnippetFilter,
+    DistrictFilter,
+    Admin2Filter,
+    EventFilter,
+    EventSnippetFilter,
+    SituationReportFilter,
+    AppealFilter,
+    AppealHistoryFilter,
+    AppealDocumentFilter,
+    FieldReportFilter,
+    GoHistoricalFilter
+)
 
 from .logger import logger
 
@@ -182,46 +201,6 @@ class RegionViewset(viewsets.ReadOnlyModelViewSet):
         return RegionRelationSerializer
 
 
-class CountryFilter(filters.FilterSet):
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-    record_type = filters.NumberFilter(field_name="record_type", lookup_expr="exact")
-    is_independent = filters.BooleanFilter(
-        field_name="independent",
-        label="is_independent",
-        lookup_expr="exact"
-    )
-    is_deprecated = filters.BooleanFilter(
-        field_name="is_deprecated",
-        label="is_deprecated",
-        lookup_expr="exact"
-    )
-    is_nationalsociety = filters.BooleanFilter(
-        label="is_nationalsociety",
-        method="filter_national_society"
-    )
-
-    class Meta:
-        model = Country
-        fields = {
-            "id": ("exact", "in"),
-            "region": ("exact", "in"),
-            "record_type": ("exact", "in"),
-        }
-
-    def filter_national_society(self, qs, name, value):
-        if not value:
-            return qs
-        return qs.filter(
-            (
-                models.Q(independent=True) &
-                models.Q(society_name__isnull=False)
-            ) |
-            (
-                (models.Q(name__icontains="RC")) | models.Q(iso="BX")
-            )
-        )
-
-
 class CountryViewset(viewsets.ReadOnlyModelViewSet):
     queryset = Country.objects.filter(is_deprecated=False).annotate(
         has_country_plan=models.Exists(CountryPlan.objects.filter(country=OuterRef("pk"), is_publish=True))
@@ -263,35 +242,11 @@ class CountryViewset(viewsets.ReadOnlyModelViewSet):
         raise Http404
 
 
-class CountryFilterRMD(filters.FilterSet):
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-
-    class Meta:
-        model = Country
-        fields = {
-            "id": ("exact", "in"),
-            "region": ("exact", "in"),
-            "record_type": ("exact", "in"),
-        }
-
-
 class CountryRMDViewset(viewsets.ReadOnlyModelViewSet):
     queryset = Country.objects.filter(is_deprecated=False).filter(iso3__isnull=False).exclude(iso3="")
     filterset_class = CountryFilterRMD
     search_fields = ("name",)
     serializer_class = CountrySerializerRMD
-
-
-class DistrictRMDFilter(filters.FilterSet):
-    class Meta:
-        model = District
-        fields = {
-            "id": ("exact", "in"),
-            "country": ("exact", "in"),
-            "country__iso3": ("exact", "in"),
-            "country__name": ("exact", "in"),
-            "name": ("exact", "in"),
-        }
 
 
 class DistrictRMDViewset(viewsets.ReadOnlyModelViewSet):
@@ -304,14 +259,6 @@ class DistrictRMDViewset(viewsets.ReadOnlyModelViewSet):
     serializer_class = DistrictSerializerRMD
 
 
-class RegionKeyFigureFilter(filters.FilterSet):
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-
-    class Meta:
-        model = RegionKeyFigure
-        fields = ("region",)
-
-
 class RegionKeyFigureViewset(ReadOnlyVisibilityViewset):
     authentication_classes = (TokenAuthentication,)
     serializer_class = RegionKeyFigureSerializer
@@ -319,27 +266,11 @@ class RegionKeyFigureViewset(ReadOnlyVisibilityViewset):
     visibility_model_class = RegionKeyFigure
 
 
-class CountryKeyFigureFilter(filters.FilterSet):
-    country = filters.NumberFilter(field_name="country", lookup_expr="exact")
-
-    class Meta:
-        model = CountryKeyFigure
-        fields = ("country",)
-
-
 class CountryKeyFigureViewset(ReadOnlyVisibilityViewset):
     authentication_classes = (TokenAuthentication,)
     serializer_class = CountryKeyFigureSerializer
     filterset_class = CountryKeyFigureFilter
     visibility_model_class = CountryKeyFigure
-
-
-class RegionSnippetFilter(filters.FilterSet):
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-
-    class Meta:
-        model = RegionSnippet
-        fields = ("region",)
 
 
 class RegionSnippetViewset(ReadOnlyVisibilityViewset):
@@ -354,14 +285,6 @@ class RegionSnippetViewset(ReadOnlyVisibilityViewset):
         return RegionSnippetSerializer
 
 
-class CountrySnippetFilter(filters.FilterSet):
-    country = filters.NumberFilter(field_name="country", lookup_expr="exact")
-
-    class Meta:
-        model = CountrySnippet
-        fields = ("country",)
-
-
 class CountrySnippetViewset(ReadOnlyVisibilityViewset):
     authentication_classes = (TokenAuthentication,)
     serializer_class = CountrySnippetSerializer
@@ -372,18 +295,6 @@ class CountrySnippetViewset(ReadOnlyVisibilityViewset):
         if is_tableau(self.request) is True:
             return CountrySnippetTableauSerializer
         return CountrySnippetSerializer
-
-
-class DistrictFilter(filters.FilterSet):
-    class Meta:
-        model = District
-        fields = {
-            "id": ("exact", "in"),
-            "country": ("exact", "in"),
-            "country__iso3": ("exact", "in"),
-            "country__name": ("exact", "in"),
-            "name": ("exact", "in"),
-        }
 
 
 class DistrictViewset(viewsets.ReadOnlyModelViewSet):
@@ -401,18 +312,6 @@ class DistrictViewset(viewsets.ReadOnlyModelViewSet):
             return DistrictSerializer
 
 
-class Admin2Filter(filters.FilterSet):
-    class Meta:
-        model = Admin2
-        fields = {
-            "id": ("exact", "in"),
-            "admin1": ("exact", "in"),
-            "admin1__country": ("exact", "in"),
-            "admin1__country__iso3": ("exact", "in"),
-            "name": ("exact", "in"),
-        }
-
-
 class Admin2Viewset(viewsets.ReadOnlyModelViewSet):
     filterset_class = Admin2Filter
     search_fields = ("name", "district__name", "district__country__name")
@@ -425,26 +324,6 @@ class Admin2Viewset(viewsets.ReadOnlyModelViewSet):
             .filter(admin1__is_deprecated=False)
             .filter(is_deprecated=False)
         )
-
-
-class EventFilter(filters.FilterSet):
-    dtype = filters.NumberFilter(field_name="dtype", lookup_expr="exact")
-    is_featured = filters.BooleanFilter(field_name="is_featured")
-    is_featured_region = filters.BooleanFilter(field_name="is_featured_region")
-    countries__in = ListFilter(field_name="countries__id")
-    regions__in = ListFilter(field_name="regions__id")
-    id = filters.NumberFilter(field_name="id", lookup_expr="exact")
-    auto_generated_source = filters.ChoiceFilter(
-        label="Auto generated source choices",
-        choices=[(v, v) for v in SOURCES.values()],
-    )
-
-    class Meta:
-        model = Event
-        fields = {
-            "disaster_start_date": ("exact", "gt", "gte", "lt", "lte"),
-            "created_at": ("exact", "gt", "gte", "lt", "lte"),
-        }
 
 
 class EventViewset(ReadOnlyVisibilityViewset):
@@ -584,14 +463,6 @@ class EventViewset(ReadOnlyVisibilityViewset):
         return super().list(request)
 
 
-class EventSnippetFilter(filters.FilterSet):
-    event = filters.NumberFilter(field_name="event", lookup_expr="exact")
-
-    class Meta:
-        model = Snippet
-        fields = ("event",)
-
-
 class EventSnippetViewset(ReadOnlyVisibilityViewset):
     authentication_classes = (TokenAuthentication,)
     serializer_class = SnippetSerializer
@@ -604,18 +475,6 @@ class SituationReportTypeViewset(viewsets.ReadOnlyModelViewSet):
     serializer_class = SituationReportTypeSerializer
     ordering_fields = ("type",)
     search_fields = ("type",)  # for /docs
-
-
-class SituationReportFilter(filters.FilterSet):
-    event = filters.NumberFilter(field_name="event", lookup_expr="exact")
-    type = filters.NumberFilter(field_name="type", lookup_expr="exact")
-
-    class Meta:
-        model = SituationReport
-        fields = {
-            "name": ("exact",),
-            "created_at": ("exact", "gt", "gte", "lt", "lte"),
-        }
 
 
 class SituationReportViewset(ReadOnlyVisibilityViewset):
@@ -636,47 +495,6 @@ class SituationReportViewset(ReadOnlyVisibilityViewset):
         if is_tableau(self.request) is True:
             return SituationReportTableauSerializer
         return SituationReportSerializer
-
-
-class AppealFilter(filters.FilterSet):
-    atype = filters.NumberFilter(field_name="atype", lookup_expr="exact")
-    dtype = filters.NumberFilter(field_name="dtype", lookup_expr="exact")
-    country = filters.NumberFilter(field_name="country", lookup_expr="exact")
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-    code = filters.CharFilter(field_name="code", lookup_expr="exact")
-    status = filters.NumberFilter(field_name="status", lookup_expr="exact")
-    id = filters.NumberFilter(field_name="id", lookup_expr="exact")
-
-    class Meta:
-        model = Appeal
-        fields = {
-            "start_date": ("exact", "gt", "gte", "lt", "lte"),
-            "end_date": ("exact", "gt", "gte", "lt", "lte"),
-        }
-
-
-class AppealHistoryFilter(filters.FilterSet):
-    atype = filters.NumberFilter(field_name="atype", lookup_expr="exact")
-    dtype = filters.NumberFilter(field_name="dtype", lookup_expr="exact")
-    country = filters.NumberFilter(field_name="country", lookup_expr="exact")
-    region = filters.NumberFilter(field_name="region", lookup_expr="exact")
-    code = filters.CharFilter(field_name="code", lookup_expr="exact")
-    status = filters.NumberFilter(field_name="status", lookup_expr="exact")
-    # Do not use, misleading: id = filters.NumberFilter(field_name='id', lookup_expr='exact')
-    appeal_id = filters.NumberFilter(
-        field_name="appeal_id", lookup_expr="exact", help_text="Use this (or code) for appeal identification."
-    )
-
-    class Meta:
-        model = AppealHistory
-        fields = {
-            "start_date": ("exact", "gt", "gte", "lt", "lte"),
-            "end_date": ("exact", "gt", "gte", "lt", "lte"),
-            "valid_from": ("exact", "gt", "gte", "lt", "lte"),
-            "valid_to": ("exact", "gt", "gte", "lt", "lte"),
-            "appeal__real_data_update": ("exact", "gt", "gte", "lt", "lte"),
-            "country__iso3": ("exact",),
-        }
 
 
 # Instead of viewsets.ReadOnlyModelViewSet:
@@ -747,18 +565,6 @@ class AppealViewset(mixins.ListModelMixin, viewsets.GenericViewSet):
 #        return Response(self.remove_unconfirmed_event(serializer.data))
 
 
-class AppealDocumentFilter(filters.FilterSet):
-    appeal = filters.NumberFilter(field_name="appeal", lookup_expr="exact")
-    appeal__in = ListFilter(field_name="appeal__id")
-
-    class Meta:
-        model = AppealDocument
-        fields = {
-            "name": ("exact",),
-            "created_at": ("exact", "gt", "gte", "lt", "lte"),
-        }
-
-
 class AppealDocumentViewset(viewsets.ReadOnlyModelViewSet):
     queryset = AppealDocument.objects.all()
     ordering_fields = (
@@ -800,21 +606,7 @@ class UserViewset(viewsets.ModelViewSet):
         return Response(self.get_serializer_class()(request.user).data)
 
 
-class FieldReportFilter(filters.FilterSet):
-    dtype = filters.NumberFilter(field_name="dtype", lookup_expr="exact")
-    user = filters.NumberFilter(field_name="user", lookup_expr="exact")
-    countries__in = ListFilter(field_name="countries__id")
-    regions__in = ListFilter(field_name="regions__id")
-    id = filters.NumberFilter(field_name="id", lookup_expr="exact")
-    is_covid_report = filters.BooleanFilter(field_name="is_covid_report")
-    summary = filters.CharFilter(field_name="summary", lookup_expr="icontains")
 
-    class Meta:
-        model = FieldReport
-        fields = {
-            "created_at": ("exact", "gt", "gte", "lt", "lte"),
-            "updated_at": ("exact", "gt", "gte", "lt", "lte"),
-        }
 
 
 @extend_schema_view(
@@ -1204,16 +996,6 @@ class MainContactViewset(viewsets.ReadOnlyModelViewSet):
 class NSLinksViewset(viewsets.ReadOnlyModelViewSet):
     serializer_class = NsSerializer
     queryset = Country.objects.filter(url_ifrc__contains="/").order_by("url_ifrc")
-
-
-class GoHistoricalFilter(filters.FilterSet):
-    countries = filters.ModelMultipleChoiceFilter(field_name="countries", queryset=Country.objects.all())
-    iso3 = filters.CharFilter(field_name="countries__iso3", lookup_expr="icontains")
-    region = filters.NumberFilter(field_name="countries__region", lookup_expr="exact")
-
-    class Meta:
-        model = Event
-        fields = ()
 
 
 class GoHistoricalViewSet(viewsets.ReadOnlyModelViewSet):
