+++ deployments/serializers.py
@@ -551,9 +551,6 @@ class EmergencyProjectActivityLocationSerializer(ModelSerializer):
 class EmergencyProjectOptionsSerializer(serializers.Serializer):
     sectors = EmergencyProjectActivitySectorSerializer(read_only=True, many=True)
     actions = EmergencyProjectActivityActionSerializer(read_only=True, many=True)
-    activity_leads = CharKeyValueSerializer(read_only=True, many=True)
-    activity_status = CharKeyValueSerializer(read_only=True, many=True)
-    activity_people_households = CharKeyValueSerializer(read_only=True, many=True)
 
 
 class EmergencyProjectActivitySerializer(
@@ -614,12 +611,15 @@ class EmergencyProjectSerializer(
 
     class Meta:
         model = EmergencyProject
-        fields = ('id', 'created_by_details', 'modified_by_details', 'reporting_ns_details',
-                  'deployed_eru_details', 'districts_details', 'activities', 'event_details',
-                  'activity_lead_display', 'status_display', 'country_details', 'visibility_display',
-                  'title', 'activity_lead', 'reporting_ns', 'event', 'country', 'districts', 'status',
-                  'created_at', 'modified_at', 'start_date', 'end_date', 'admin2', 'admin2_details',
-                  )  # '__all__' | Both X_details + X fields are needed: outgoing + incoming data.
+        fields = (
+            'id', 'created_by_details', 'modified_by_details', 'reporting_ns_details',
+            'deployed_eru_details', 'districts_details', 'activities', 'event_details',
+            'activity_lead_display', 'status_display', 'country_details', 'visibility_display',
+            'title', 'activity_lead', 'reporting_ns', 'event', 'country', 'districts', 'status',
+            'created_at', 'modified_at', 'start_date', 'end_date', 'admin2', 'admin2_details',
+            'reporting_ns_contact_name', 'reporting_ns_contact_role',
+            'reporting_ns_contact_email', 'deployed_eru',
+        )  # '__all__' | Both X_details + X fields are needed: outgoing + incoming data.
         read_only_fields = (
             'created_by',
             'created_at',
